Task. 4. Spring IoC

+++ liquibase/changelogs/homework-2/sql/05.create-comments-table.sql
@@ -1,11 +1,10 @@
 CREATE TABLE comments
 (
-    id   SERIAL PRIMARY KEY,
-    task_id      INT,
-    employee_id  INT,
+    id           SERIAL PRIMARY KEY,
+    task_id      BIGINT,
+    employee_id  BIGINT,
     comment_text TEXT NOT NULL,
     created_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (task_id) REFERENCES tasks (task_id),
     FOREIGN KEY (employee_id) REFERENCES employees (employee_id)
 );
-
